move code into src
