+++ merged-packages/json-rpc-engine/package.json
@@ -2,7 +2,7 @@
   "name": "json-rpc-engine",
   "version": "2.2.0",
   "description": "a tool for processing JSON RPC",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "npm run test-node && npm run test-browser",
     "test-node": "node test/index.js",
